add icons into the world overlay
